Image links extracting from HTML pages.
Images are extract from img-src tags. Links are stored in solr in a new
field in schema.xml (links_images). The feature is optional.
Config: warc.index.extract.linked.images.
Both the schema.xml (if enabled) and config.conf must define this new
field. 
In the previous version image link extraction was dead code. And if
enabled it would store image in the same fields as the normal links (a
href). Now it goes to its own field (if enabled)